# Optimize DeflateStream.CopyToAsync