--- a/src/System.IO.Compression/src/System/IO/Compression/DeflateZLib/Inflater.cs
+++ b/src/System.IO.Compression/src/System/IO/Compression/DeflateZLib/Inflater.cs
@@ -13,13 +13,15 @@ namespace System.IO.Compression
     /// </summary>
     internal sealed class Inflater : IDisposable
     {
+        private const int MinWindowBits = -15;              // WindowBits must be between -8..-15 to ignore the header, 8..15 for
+        private const int MaxWindowBits = 47;               // zlib headers, 24..31 for GZip headers, or 40..47 for either Zlib or GZip
+
         private bool _finished;                             // Whether the end of the stream has been reached
         private bool _isDisposed;                           // Prevents multiple disposals
         private ZLibNative.ZLibStreamHandle _zlibStream;    // The handle to the primary underlying zlib stream
         private GCHandle _inputBufferHandle;                // The handle to the buffer that provides input to _zlibStream
-        private readonly object _syncLock = new object();   // Used to make writing to unmanaged structures atomic 
-        private const int minWindowBits = -15;              // WindowBits must be between -8..-15 to ignore the header, 8..15 for
-        private const int maxWindowBits = 47;               // zlib headers, 24..31 for GZip headers, or 40..47 for either Zlib or GZip
+
+        private object SyncLock => this;                   // Used to make writing to unmanaged structures atomic 
 
         #region Exposed Members
 
@@ -28,7 +30,7 @@ namespace System.IO.Compression
         /// </summary>
         internal Inflater(int windowBits)
         {
-            Debug.Assert(windowBits >= minWindowBits && windowBits <= maxWindowBits);
+            Debug.Assert(windowBits >= MinWindowBits && windowBits <= MaxWindowBits);
             _finished = false;
             _isDisposed = false;
             InflateInit(windowBits);
@@ -110,7 +112,7 @@ namespace System.IO.Compression
             if (0 == count)
                 return;
 
-            lock (_syncLock)
+            lock (SyncLock)
             {
                 _inputBufferHandle = GCHandle.Alloc(inputBuffer, GCHandleType.Pinned);
                 _zlibStream.NextIn = _inputBufferHandle.AddrOfPinnedObject() + startIndex;
@@ -189,7 +191,7 @@ namespace System.IO.Compression
         /// </summary>
         private unsafe ZLibNative.ErrorCode ReadInflateOutput(byte* bufPtr, int length, ZLibNative.FlushCode flushCode, out int bytesRead)
         {
-            lock (_syncLock)
+            lock (SyncLock)
             {
                 _zlibStream.NextOut = (IntPtr)bufPtr;
                 _zlibStream.AvailOut = (uint)length;
@@ -246,7 +248,7 @@ namespace System.IO.Compression
         {
             Debug.Assert(_inputBufferHandle.IsAllocated);
 
-            lock (_syncLock)
+            lock (SyncLock)
             {
                 _zlibStream.AvailIn = 0;
                 _zlibStream.NextIn = ZLibNative.ZNullPtr;