Today, CopyToAsync on DeflateStream is the basic read/write loop.  However, all CopyToAsync really needs to do on DeflateStream is transform the source data (via a synchronous operation) and pass it along to the destination.  This means that we can improve its implementation by effectively having the source stream CopyToAsync to the destination stream, albeit with that transformation added in the middle.  Doing this has a three key benefits:
- It eliminates the read operations and the associated copy, allowing the source to effectively push directly to the inflater inside DeflateStream.
- It takes advantage of any optimizations the source stream has in its CopyToAsync implementation, e.g. MemoryStream simply passing its whole internal buffer, NetworkStream using a SocketAsyncEventArgs (https://github.com/dotnet/corefx/pull/12664), FileStream reusing its async state (https://github.com/dotnet/corefx/pull/11569), etc.
- CopyToAsync is typically used when no other reading has been done from the stream.  Since this change causes CopyToAsync to bypass the internal buffer used by DeflateStream before the inflater itself, we can avoid allocating that buffer at all if CopyToAsync is used by itself.  This avoids an 8K allocation per DeflateStream instance, and even if we switch to pulling this buffer from ArrayPool, it still avoids needing to rent/return with the ArrayPool in these cases.

(I also noticed an unnecessary allocation in Inflater, which I removed.)

The net effect of this change is a sizeable boost on throughput for many scenarios (I didn't find any scenarios where it regressed).  Here are a few examples.  In this table, "MS" is a memory stream, so "MS to MS" is using DeflateStream created around a compressed MemoryStream and using CopyToAsync on the DeflateStream to decompress it to another MemoryStream.  "CMS" is a "CustomMemoryStream", which is simply a type that derives from MemoryStream and doesn't add or override any logic; this derivation defeats some the CopyToAsync optimization in MemoryStream, so this is essentially the worst case for this DeflateStream optimization.  And "FS" is FileStream, where a "(true)" suffix means useAsync==true (in which case FileStream has its own CopyToAsync implementation) and a "(false)" suffix means useAsync==false (in which case it essentially delegates back to the base Stream's CopyToAsync). "Len" is the length of the original random input in bytes, and "Iters" is the number of iterations (the number of times positions were reset and CopyToAsync was called).

| Test                                              | Before (s) | After (s)|
|---------------------------------------------------|-----------|-----------|
| MS to MS (Len=1024     Iters=1024000)             |  2.767149 | 1.601566  |
| MS to MS (Len=1048576  Iters=1000)                |  0.419615 | 0.139993  |
| MS to MS (Len=10485760 Iters=100)                 |  0.358287 | 0.223764  |
| CMS to CMS (Len=1024     Iters=1024000)           |  2.832110 | 1.963891  |
| CMS to CMS (Len=1048576  Iters=1000)              |  0.271432 | 0.202536  |
| CMS to CMS (Len=10485760 Iters=100)               |  0.347356 | 0.316338  |
| FS(True) to FS(True) (Len=1024     Iters=102400)  | 10.315941 | 6.909713  |
| FS(True) to FS(True) (Len=1048576  Iters=100)     |  0.940410 | 0.124353  |
| FS(True) to FS(True) (Len=10485760 Iters=10)      |  1.000786 | 0.152991  |
| FS(True) to FS(False) (Len=1024     Iters=102400) |  9.936322 | 6.967522  |
| FS(True) to FS(False) (Len=1048576  Iters=100)    |  0.705946 | 0.085895  |
| FS(True) to FS(False) (Len=10485760 Iters=10)     |  0.925196 | 0.159865  |
| FS(False) toFS (True) (Len=1024     Iters=102400) |  8.630263 | 6.699364  |
| FS(False) toFS (True) (Len=1048576  Iters=100)    |  0.880577 | 0.116051  |
| FS(False) toFS (True) (Len=10485760 Iters=10)     |  0.862978 | 0.142838  |
| FS(False) toFS (False) (Len=1024     Iters=102400)|  5.260478 | 4.388572  |
| FS(False) toFS (False) (Len=1048576  Iters=100)   |  0.346608 | 0.078440  |
| FS(False) toFS (False) (Len=10485760 Iters=10)    |  0.432393 | 0.157312  | 

Fixes https://github.com/dotnet/corefx/issues/11571
Contributes to https://github.com/dotnet/corefx/issues/12549
cc: @ianhays, @JeremyKuhne, @geoffkizer, @benaadams, @davidfowl 

(@ianhays, please pay careful attention to my use of the inflater.  I essentially copied what was in ReadAsync and refactored it, and I want to make sure I'm using it correctly / didn't misunderstand anything about its behavior.)