@geoffkizer 2016-10-31
Overall this looks great; some minor nits above.